Add support for custom editorLangId array
Following discussion in #46, this commit contributes a
markdownShortcuts.languages configuration option. Users can provide an
array of editorLangId strings for which the extension will be enabled.

The original implementation idea was to save the RegExp object or string
in the extension context and use that within package.json:

"when": "editorLangId =~ markdownShortcuts:languages"

But the =~ operator only seems to work with RegExp literals.

The alternative solution provided in this commit is to save the boolean
result of that RegExp test in markdownShortcuts:enabled and use that
instead:

"when": "markdownShortcuts:enabled"

The downside is that this boolean needs to be updated. Adding listeners
for the following events seems to cover all cases:

- onDidChangeConfiguration to listen for settings.json changes
- onDidChangeActiveTextEditor to update enabled state when the editor in
  focus changes
- onDidOpenTextDocument to update enabled state when the language id
  assigned to a document changes

Currently:

- If markdownShortcuts.languages = [], the extension is still enabled on
  Markdown files.
- If markdownShortcuts.languages = ["plaintext"], the extension will be
  enabled on Plain Text files but not on Markdown files.
- Finally, markdownShortcuts.languages = ["markdown", "plaintext"] works
  as expected.

Closes #46.